api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
slowapi==0.1.9
python-dotenv==1.0.1
ecdsa==0.19.0
